+++ library/src/nvidia_detail/hipsolver_sparse.cpp
@@ -149,7 +149,7 @@ catch(...)
     return exception2hip_status();
 }
 
-hipsolverStatus_t hipsolverSpCcsrlsvchol(hipsolverSpHandle_t       handle,
+/*hipsolverStatus_t hipsolverSpCcsrlsvchol(hipsolverSpHandle_t       handle,
                                          int                       n,
                                          int                       nnzA,
                                          const hipsparseMatDescr_t descrA,
@@ -221,7 +221,7 @@ try
 catch(...)
 {
     return exception2hip_status();
-}
+}*/
 
 hipsolverStatus_t hipsolverSpScsrlsvcholHost(hipsolverSpHandle_t       handle,
                                              int                       n,
@@ -297,7 +297,7 @@ catch(...)
     return exception2hip_status();
 }
 
-hipsolverStatus_t hipsolverSpCcsrlsvcholHost(hipsolverSpHandle_t       handle,
+/*hipsolverStatus_t hipsolverSpCcsrlsvcholHost(hipsolverSpHandle_t       handle,
                                              int                       n,
                                              int                       nnzA,
                                              const hipsparseMatDescr_t descrA,
@@ -369,6 +369,6 @@ try
 catch(...)
 {
     return exception2hip_status();
-}
+}*/
 
 } //extern C