+++ CHANGELOG.md
@@ -7,8 +7,8 @@ Full documentation for hipSOLVER is available at [hipsolver.readthedocs.io](http
 - Added compatibility API with hipsolverSp prefix
 - Added compatibility-only functions
   - csrlsvchol
-    - hipsolverSpScsrlsvcholHost, hipsolverSpDcsrlsvcholHost, hipsolverSpCcsrlsvcholHost, hipsolverSpZcsrlsvcholHost
-    - hipsolverSpScsrlsvchol, hipsolverSpDcsrlsvchol, hipsolverSpCcsrlsvchol, hipsolverSpZcsrlsvchol
+    - hipsolverSpScsrlsvcholHost, hipsolverSpDcsrlsvcholHost
+    - hipsolverSpScsrlsvchol, hipsolverSpDcsrlsvchol
 - Added hipBLAS as an optional dependency to hipsolver-test. Use the `BUILD_HIPBLAS_TESTS` CMake option to test compatibility between hipSOLVER and hipBLAS.
 - Added hipSPARSE as an optional dependency to hipsolver-test. Tests for hipsolverSp will only be run if built with hipSPARSE installed on the system.